Custom Display Images
Self explanatory because I'm so good at naming things. ImageImporterAndScaler is the script used to import images from the file browser. It relies on the TextureScale and Converter to work it's magic so like, don't mess with those too much.
TextureScaler is dead though, hopefully no one was using that.